creative/app/requirements.txt
Flask==1.1.2
Flask-WTF==0.14.3
google-cloud-firestore==2.0.2
Flask-Bootstrap==3.3.7.1
